+++ packages/protocol/src/peer/ControllerCommissioningFlow.ts
@@ -1379,7 +1379,6 @@ export class ControllerCommissioningFlow {
                     command: "updateFabricLabel",
                     fields: {
                         label: this.fabric.label,
-                        fabricIndex,
                     },
                 }),
             );